changed: improve visualization handling.
- introduce a manager for instances of visualisation addons
- use this to detach a visualisation's lifecycle from a visualisation
  gui control's life cycle
- introduce multiple slots for visualization - allowing multiple
  displayed at the same time. this is exposed to skinners through the
  subtags 'slot' (just some number) and 'id' (the add-on id for the
  visualization).

add-ons needs to be updated, as the API changes. we have to change the Start()
method to pass screen info since this is now called once per gui control
instance. Likewise Stop() is called when the gui control is killed.

Create() is called at the beginning of the life cycle, and Create()
at the end. These are called exactly once.