RefreshChangeDelay: Introduce Auto setting
Use a negative value of -1 to indicate Auto mode. This integrates in a backward
compatible way as for now only positive values triggered the logic. For Android
things are special as here a mix of trust to get the event and checking
explicitely with a 2s default timer is used. This new Auto setting will
allow Android to wait until the Audio Hotplug event is coming, with the downside
that it might never come. Therefore we keep Off as the default.